Add considering Ram in the process and optimization + Fixes

+++ PcbDesignSimuModeling.Core/Models/Resources/Ram/RamRequest.cs
@@ -4,7 +4,7 @@ public class RamRequest : ResourceRequest<Ram>
 {
     private readonly double _amount;
 
-    public RamRequest(int requestId, int amount) : base(requestId) => _amount = amount;
+    public RamRequest(int requestId, double amount) : base(requestId) => _amount = amount;
 
     protected override bool TryGetResourceBody(Ram potentialResource) =>
         potentialResource.TryGetResource(RequestId, _amount);
